+++ docs/changelog.md
@@ -1,9 +1,13 @@
-#### 2022/08/09
+#### 2022/08/11
+
+* Add UoT option for [Shadowsocks](/configuration/outbound/shadowsocks) outbound, UoT support for all inbounds.
+
+#### 2022/08/10
 
 * Add full-featured [Naive](/configuration/inbound/naive) inbound
 * Fix default dns server option [#9] by iKirby
 
-#### 2022/08/08
+#### 2022/08/09
 
 No changelog before.